# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DemoResult": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "alert_ids": {
            "items": {
              "type": "string"
            },
            "title": "Alert Ids",
            "type": "array"
          },
          "credential_id": {
            "title": "Credential Id",
            "type": "string"
          },
          "destructive_attempt": {
            "title": "Destructive Attempt",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "sandboxed": {
            "default": true,
            "title": "Sandboxed",
            "type": "boolean"
          },
          "scanner_warning": {
            "title": "Scanner Warning",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "session_id",
          "agent_id",
          "credential_id",
          "finding_id",
          "alert_ids",
          "scanner_warning",
          "destructive_attempt"
        ],
        "title": "DemoResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicDemoRequest": {
        "properties": {
          "visitor_ref": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Visitor Ref",
            "type": "string"
          }
        },
        "required": [
          "visitor_ref"
        ],
        "title": "PublicDemoRequest",
        "type": "object"
      },
      "PublicDemoResult": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "alert_ids": {
            "items": {
              "type": "string"
            },
            "title": "Alert Ids",
            "type": "array"
          },
          "credential_id": {
            "title": "Credential Id",
            "type": "string"
          },
          "destructive_attempt": {
            "title": "Destructive Attempt",
            "type": "string"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "sandboxed": {
            "default": true,
            "title": "Sandboxed",
            "type": "boolean"
          },
          "scanner_warning": {
            "title": "Scanner Warning",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "share_path": {
            "title": "Share Path",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "session_id",
          "agent_id",
          "credential_id",
          "finding_id",
          "alert_ids",
          "scanner_warning",
          "destructive_attempt",
          "share_path",
          "expires_at"
        ],
        "title": "PublicDemoResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Leaflyst API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/demo/public/run": {
      "post": {
        "operationId": "run_public_demo_v1_demo_public_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicDemoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDemoResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Public Demo"
      }
    },
    "/v1/demo/run": {
      "post": {
        "operationId": "run_demo_v1_demo_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Demo"
      }
    },
...
  }
}